requirements.txt
fastapi==0.74.0
sqlalchemy==1.4.31
"uvicorn[standard]"
databases==0.5.5
psycopg2==2.8.6
alembic==1.7.5
python-dotenv